+++ CHANGELOG.md
@@ -81,6 +81,8 @@
 - Changes to xSQLServerRSConfig
   - No longer returns a null value from Test-TargetResource when Reporting
     Services has not been initialized ([issue #822](https://github.com/PowerShell/xSQLServer/issues/822)).
+  - Fixed so that when two Reporting Services are installed for the same major
+    version the resource does not throw an error ([issue #819](https://github.com/PowerShell/xSQLServer/issues/819)).
 
 ## 8.1.0.0